Move TEC-1G helpers into platform ownership
